account and function updates
moved the account update data in to one file for organisation purposes.

Added feature where a user is unable to create an account or update an account with an email that already exists in database.

Added more system dialogues.

Changed the way the placeholder image is handled (suggested by Vincent. Thank you!)