changed some css and continued working on search
Have one code that SELECTS everything LIKE what the user has searched for, but not connected to user_id.

Have another code that SELECTS everything LIKE what the user searched for except whats in "description" for some reason, but connected to the right user_id.

Cant figure out the problem.